feat: upload file to source

+++ app/UseCases/Project/UpdateProjectUseCase.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\UseCases\Project;
 
+use App\Helpers\FileManager;
 use App\Models\Project;
 use App\Repositories\GalleryRepository;
 use App\Repositories\MediaItemRepository;
@@ -13,6 +14,7 @@ use Common\App\Enums\WebVisibility;
 use Exception;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Storage;
 
 /**
  * The use case class for updating a project.
@@ -37,11 +39,26 @@ class UpdateProjectUseCase
         string $summaryVi,
         bool $isHighlight,
         WebVisibility $webVisibility,
-        array $thumbnail,
+        string $thumbnailFileUrl,
         array $galleries
     ): array {
         try {
+            $deletedTmpFilePaths = [];
+            $deletedPublicFilePaths = [];
+
             DB::beginTransaction();
+
+            $originalThumbnailFilePath = $project->thumbnail_file_path;
+            $newThumbnailFilePath = $originalThumbnailFilePath;
+            $originalGalleryIds = $project->galleries->pluck('id')->all();
+
+            // Case: Change thumbnail.
+            if (str_starts_with($thumbnailFileUrl, config('filesystems.disks.tmp.url'))) {
+                $thumbnailFileName = basename($thumbnailFileUrl);
+                $newThumbnailFilePath = FileManager::moveTmpToPublic($thumbnailFileName, 'projects/thumbnails');
+                $deletedTmpFilePaths[] = $thumbnailFileName;
+                $deletedPublicFilePaths[] = $originalThumbnailFilePath;
+            }
 
             // Update project.
             $this->projectRepository->update(
@@ -54,25 +71,29 @@ class UpdateProjectUseCase
                 $summaryEn,
                 $summaryVi,
                 $isHighlight,
-                $thumbnail['file_id'],
-                $thumbnail['file_name'],
-                $thumbnail['file_mime_type'],
+                $newThumbnailFilePath,
                 $webVisibility
             );
 
             // Case: Add new galleries.
-            $newGalleryIds = array_diff(array_column($galleries, 'id'), $project->galleries->pluck('id')->toArray());
+            $newGalleryIds = array_diff(array_column($galleries, 'id'), $originalGalleryIds);
             $newGalleries = array_filter($galleries, fn ($gallery) => in_array($gallery['id'], $newGalleryIds));
 
             foreach ($newGalleries as $gallery) {
                 $newGallery = $this->galleryRepository->create($project->id, $gallery['caption']);
 
                 foreach ($gallery['media_items'] as $mediaItem) {
+                    $mediaItemFileUrl = $mediaItem['file_url'];
+                    $mediaItemFileName = basename($mediaItemFileUrl);
+                    $mediaItemFilePath = FileManager::moveTmpToPublic(
+                        $mediaItemFileName,
+                        "projects/{$project->id}/{$newGallery->id}"
+                    );
+                    $deletedTmpFilePaths[] = $mediaItemFileName;
+
                     $this->mediaItemRepository->create(
                         $newGallery->id,
-                        $mediaItem['file_id'],
-                        $mediaItem['file_name'],
-                        $mediaItem['file_mime_type'],
+                        $mediaItemFilePath,
                         MediaFrame::from($mediaItem['frame']),
                         $mediaItem['is_banner']
                     );
@@ -80,62 +101,66 @@ class UpdateProjectUseCase
             }
 
             // Case: Delete galleries.
-            $deletedGalleryIds = array_diff($project->galleries->pluck('id')->toArray(), array_column($galleries, 'id'));
+            $deletedGalleryIds = array_diff($originalGalleryIds, array_column($galleries, 'id'));
             $deletedGalleries = $project->galleries->whereIn('id', $deletedGalleryIds);
-            $deletedMediaItems = $deletedGalleries->pluck('mediaItems')->flatten();
-            $deletedMediaItemIds = $deletedMediaItems->pluck('id')->toArray();
 
-            $this->mediaItemRepository->bulkDelete($deletedMediaItemIds);
-            $this->galleryRepository->bulkDelete($deletedGalleryIds);
+            if (filled($deletedGalleries)) {
+                $deletedMediaItems = $deletedGalleries->pluck('mediaItems')->flatten();
+                $deletedMediaItemFilePaths = $deletedMediaItems->pluck('file_path')->toArray();
+                $deletedMediaItemIds = $deletedMediaItems->pluck('id')->toArray();
+
+                array_push($deletedPublicFilePaths, ...$deletedMediaItemFilePaths);
+
+                $this->mediaItemRepository->bulkDelete($deletedMediaItemIds);
+                $this->galleryRepository->bulkDelete($deletedGalleryIds);
+            }
 
             // Case: Update galleries.
-            $updatedGalleryIds = array_intersect(
-                array_column($galleries, 'id'),
-                $project->galleries->pluck('id')->toArray()
-            );
+            $updatedGalleryIds = array_intersect(array_column($galleries, 'id'), $originalGalleryIds);
             $updatedGalleries = array_filter($galleries, fn ($gallery) => in_array($gallery['id'], $updatedGalleryIds));
 
             foreach ($updatedGalleries as $gallery) {
                 $galleryId = (int) $gallery['id'];
                 $originalMediaItems = $project->galleries->firstWhere('id', $galleryId)->mediaItems;
+                $originalMediaItemIds = $originalMediaItems->pluck('id')->all();
 
                 $this->galleryRepository->update($galleryId, $gallery['caption']);
 
                 // Case: Add new media items.
-                $newMediaItemIds = array_diff(
-                    array_column($gallery['media_items'], 'id'),
-                    $originalMediaItems->pluck('id')->toArray()
-                );
+                $newMediaItemIds = array_diff(array_column($gallery['media_items'], 'id'), $originalMediaItemIds);
                 $newMediaItems = array_filter(
                     $gallery['media_items'],
                     fn ($mediaItem) => in_array($mediaItem['id'], $newMediaItemIds)
                 );
 
                 foreach ($newMediaItems as $mediaItem) {
+                    $mediaItemFileUrl = $mediaItem['file_url'];
+                    $mediaItemFileName = basename($mediaItemFileUrl);
+                    $mediaItemFilePath = FileManager::moveTmpToPublic(
+                        $mediaItemFileName,
+                        "projects/{$project->id}/{$galleryId}"
+                    );
+                    $deletedTmpFilePaths[] = $mediaItemFileName;
+
                     $this->mediaItemRepository->create(
                         $galleryId,
-                        $mediaItem['file_id'],
-                        $mediaItem['file_name'],
-                        $mediaItem['file_mime_type'],
+                        $mediaItemFilePath,
                         MediaFrame::from($mediaItem['frame']),
                         $mediaItem['is_banner']
                     );
                 }
 
                 // Case: Delete media items.
-                $deletedMediaItemIds = array_diff(
-                    $originalMediaItems->pluck('id')->toArray(),
-                    array_column($gallery['media_items'], 'id')
-                );
+                $deletedMediaItemIds = array_diff($originalMediaItemIds, array_column($gallery['media_items'], 'id'));
                 $deletedMediaItems = $originalMediaItems->whereIn('id', $deletedMediaItemIds);
+                $deletedMediaItemFilePaths = $deletedMediaItems->pluck('file_path')->toArray();
+
+                array_push($deletedPublicFilePaths, ...$deletedMediaItemFilePaths);
 
                 $this->mediaItemRepository->bulkDelete($deletedMediaItemIds);
 
                 // Case: Update media items.
-                $updatedMediaItemIds = array_intersect(
-                    array_column($gallery['media_items'], 'id'),
-                    $originalMediaItems->pluck('id')->toArray()
-                );
+                $updatedMediaItemIds = array_intersect(array_column($gallery['media_items'], 'id'), $originalMediaItemIds);
                 $updatedMediaItems = array_filter(
                     $gallery['media_items'],
                     fn ($mediaItem) => in_array($mediaItem['id'], $updatedMediaItemIds)
@@ -151,6 +176,9 @@ class UpdateProjectUseCase
                     );
                 }
             }
+
+            Storage::disk('tmp')->delete($deletedTmpFilePaths);
+            Storage::disk('public')->delete($deletedPublicFilePaths);
 
             DB::commit();
